changed border radius to percentage from pixels
This is cool because it changed from a square to a circle. I guess because it's a radius, it becomes a circle once it gets below a certain percentage? Twitter does this to the images users upload as their profile pic. If you click "view in browser" then you will see the full image which is actually a square. They have a smaller border-width I'm guessing 1 to 2 pixels; if any.