added fallback background color
placing a more widely accepted expression, color written out, before a less widely used expression, a variable, allows the browser to choose the correct one. Since code/scirpt always reads from top to bottom, the latest code/script is what will run. If the variable is not recognized, that is ok because bg color red is common. If bg color red were chosen first it would be accepted over the variable because it would be the last thing to be read.